Setting up a GWT framework with Specs2
Summary:
The following is a first attempt at setting up acceptance tests with Specs2
using the Given-When-Then convention.

There are multiple different ways to specify this kind of test case in
Specs2; see
https://etorreborre.github.io/specs2/guide/SPECS2-3.6/org.specs2.guide.GivenWhenThenStyle.html

The method chosen (full support) is the one I feel best combines
succinctness with readability. Specs can be grouped in individual
Scala files, and each plugin/library can have its own package
(e.g., tests for observations would be in /test/observations).

For the HttpClient, I chose to use scalaj-http:
https://github.com/scalaj/scalaj-http

It's a very simple Httpclient library, with minimal dependencies on
external libraries which does not involve messing with Futures (I'm
looking at you, Play WS). If we want to, we can switch it out later.

The controller is just a simple app to simplify developmet (you can
not use the logger in tests which is a pain).

Test Plan:
URI and client-id are hard-coded in for now. This should obviously
be fixed later.

To change the URI you're testing against, edit `val uri=...`; line 60 of HelloSpec

- Set up a functionaing METAPI, or test against https://data.met.no
- Change the client_id to a valid client_id for the API in line 48 of HelloSpec.scala (this should work correctly without manual intervention once D655 is commited).
- `activator clean test`

Activator should run through the three scenarios correctly, assuming the URI is correct and the client_id is valid.

Reviewers: #met_api, borgem

Reviewed By: #met_api, borgem

Subscribers: borgem

Maniphest Tasks: T837

Differential Revision: https://phab.met.no/D673