Split step defs across multiple packages/files
Summary:
Putting all step definitions in the same package/file doesn't scale.
Code maintenance is bound to become harder as more and more step
defs are kept in the same file.

Note how the package names for active step defs must be listed in the
'glue' array in the CucumberOptions in RunCucumber.scala. Other step defs
will not be processed.

Test Plan: Run tests according to README.md and verify that it works.

Reviewers: michaeloa

Reviewed By: michaeloa

Differential Revision: https://phab.met.no/D1892