Added framework for comparing JSON subsets and value regexps
Summary:
In Cucumber-based acceptance testing, it is sometimes useful to be able
to decide if only a subset of the JSON body is present in the
HTTP response. Also, it is sometimes useful to be able to specify
response values as regular expressions if we're interested in verifying
the form rather than the exact content (typically if the latter
varies between requests).

Test Plan:
Run tests according to README.md. Experiment with different
values in genericrequest-example1.feature and see how the test output
changes accordingly.

Reviewers: michaeloa

Reviewed By: michaeloa

Differential Revision: https://phab.met.no/D1904